Add working and tested encoder communication - Just general response implemented

+++ encoders/orbis_brd10/br10_sp.h
@@ -6,7 +6,7 @@
 typedef struct {
     bool isError;
     bool isWarning;
-    float position;
+    double position;
     char crc;
 } gen_response;
